added http responses to UserController and SecretController

+++ src/main/java/com/user/secrets/controller/SecretController.java
@@ -3,6 +3,7 @@ package com.user.secrets.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.user.secrets.dao.Secret;
 import com.user.secrets.dao.User;
+import com.user.secrets.response.SecretHTTPResponse;
 import com.user.secrets.service.SecretServiceImpl;
 import com.user.secrets.service.UserServiceImpl;
 
@@ -21,11 +23,14 @@ public class SecretController {
 
 	UserServiceImpl userServiceImpl;
 	SecretServiceImpl secretServiceImpl;
+	SecretHTTPResponse response;
 
 	@Autowired
-	public SecretController(UserServiceImpl userServiceImpl, SecretServiceImpl secretServiceImpl) {
+	public SecretController(UserServiceImpl userServiceImpl, SecretServiceImpl secretServiceImpl,
+			SecretHTTPResponse response) {
 		this.userServiceImpl = userServiceImpl;
 		this.secretServiceImpl = secretServiceImpl;
+		this.response = response;
 	}
 
 	@RequestMapping(value = "/secrets", method = RequestMethod.GET)
@@ -35,27 +40,43 @@ public class SecretController {
 	}
 
 	@RequestMapping(value = "/secret/{id}", method = RequestMethod.GET)
-	public @ResponseBody Secret getSecret(@PathVariable(value = "id") Long id) {
-		// System.out.println(.toString());
-		return secretServiceImpl.findById(id);
+	public @ResponseBody ResponseEntity<Secret> getSecret(@PathVariable(value = "id") Long id) {
+		if (ValidateSecret(id) == null) {
+			return response.notFound("secret not found: " + id);
+		} else {
+			return response.ok(secretServiceImpl.findById(id));
+		}
 	}
 
 	@RequestMapping(value = "/secret", method = RequestMethod.POST)
-	public @ResponseBody Secret saveSecret(@RequestBody Secret secret) {
+	public @ResponseBody ResponseEntity<Secret> saveSecret(@RequestBody Secret secret) {
+		if (secretServiceImpl.findById(secret.getId()) != null)
+			return response.conflict("secret already exist: " + secret.getId());
+
 		secretServiceImpl.save(secret);
-		return secret;
+		return response.created(secret);
 	}
 
 	@RequestMapping(value = "/secret/{id}", method = RequestMethod.DELETE)
-	public void deleteSecret(@PathVariable(value = "id") Long id) {
-		// System.out.println(.toString());
+	public ResponseEntity<Secret> deleteSecret(@PathVariable(value = "id") Long id) {
+		if (ValidateSecret(id) == null)
+			return response.notFound("secret not found: " + id);
+
 		secretServiceImpl.delete(id);
+		return response.ok("user deleted: " + id);
 	}
-	
+
 	@RequestMapping(value = "/secret/{id}", method = RequestMethod.PUT)
-	public void updateSecret(@PathVariable(value = "id") Long id, @RequestBody Secret secret) {
-		// System.out.println(.toString());
-		secretServiceImpl.update(id,secret);
+	public ResponseEntity<Secret> updateSecret(@PathVariable(value = "id") Long id, @RequestBody Secret secret) {
+		if (ValidateSecret(id) == null)
+			return response.notFound("secret not found: " + id);
+		secretServiceImpl.update(secret);
+		return response.ok(secret);
+
+	}
+
+	public Secret ValidateSecret(Long id) {
+		return secretServiceImpl.findById(id);
 	}
 
 }
